+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes
 
+## 0.16.0+2
+
+- fix [DataRef.compareTo]
+- test some funcs of [DatabaseTableRefs] and [DatabaseTableDict]
+
 ## 0.16.0+1
 
 - Set last Dart SDK (2.16.1) dependecy